Modify tests to respect naming convention + fix repository tests + create new frontend table to highlight the order-products relationship

+++ src/main/java/com/project/service/impl/OrderServiceImpl.java
@@ -78,6 +78,9 @@ public class OrderServiceImpl implements OrderService {
          throw new InvalidDataException(ORDER_WITH_INVALID_ID, id);
       }
       Order order = getOrderById(id);
+      if (order == null) {
+         throw new ResourceNotFoundException(NO_ORDER_FOUND);
+      }
       orderRepository.delete(order);
 
    }
